crash-reporter: convert number to int before calling abs()
BUG=chromium:347721
TEST=emerge-x86-generic crash-reporter passes.

Reviewed-on: https://chromium-review.googlesource.com/188212
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>

+++ crash_reporter/kernel_collector.cc
@@ -392,7 +392,8 @@ bool KernelCollector::FindCrashingFunction(
     return false;
   }
   if (stack_trace_timestamp != 0 &&
-      abs(stack_trace_timestamp - timestamp) > kSignatureTimestampWindow) {
+      abs(static_cast<int>(stack_trace_timestamp - timestamp))
+        > kSignatureTimestampWindow) {
     if (print_diagnostics) {
       printf("Found crashing function but not within window.\n");
     }
